node-main/deps/v8/test/inspector/heap-profiler/collect-garbage-expected.txt

Tests collectGarbage.

Running test: testCollectGarbage
WeakRef state: WeakRef is cleared after GC.
